CLI-7 Service's config string saved in Configuration

+++ CLIENTS/Data/DbContext.cs
@@ -17,13 +17,13 @@ namespace UPB.ProyectoFinal.Data
 
         public DbContext(IConfiguration config)
         {
-            ruta = config.GetSection("Location").GetSection("DirJson").Value;
-            ClientTable = new List<Client>()
-            { };
-            File.WriteAllText(ruta, Newtonsoft.Json.JsonConvert.SerializeObject(ClientTable));
-
-
-
+            ruta = config.GetConnectionString("Database");
+            if (!File.Exists(ruta))
+            {
+                ClientTable = new List<Client>()
+                { };
+                File.WriteAllText(ruta, Newtonsoft.Json.JsonConvert.SerializeObject(ClientTable));
+            }
         }
 
         public List<Client> GetAllClients()
